requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
dataclasses==0.6
Django==4.2.20
django-dotenv==1.4.2
djangorestframework==3.15.2
sqlparse==0.5.3
typing-extensions==4.13.2
